chore: Move most validation to `ProposedBatch`
