+++ crates/miden-batch-prover/src/lib.rs
@@ -6,9 +6,6 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
-mod error;
-pub use error::BatchError;
-
 mod local_batch_prover;
 pub use local_batch_prover::LocalBatchProver;
